requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# LangChain
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5

# Google Gemini
google-generativeai==0.5.4

# Vector DB
chromadb==0.5.0

# Document loaders
pypdf==4.2.0
docx2txt==0.8